OneDrive/Documents/agriculture/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.32
requests==2.32.3
pyserial==3.5

